fix(backend): (1ff2dc83a4e2ff397350b55b3ae86098702019ec)

+++ backend/src/main/java/com/onlyoffice/docspacepipedrive/security/provider/JwtAuthenticationProvider.java
@@ -78,8 +78,8 @@ public class JwtAuthenticationProvider implements AuthenticationProvider {
             throw new InvalidBearerTokenException(e.getMessage(), e);
         }
 
-        Long clientId = Long.valueOf((Integer) body.get(clientNameAttribute));
-        Long userId = Long.valueOf((Integer) body.get(userNameAttribute));
+        Long clientId = (Long) body.get(clientNameAttribute);
+        Long userId = (Long) body.get(userNameAttribute);
 
         User user;
         try {
